refactor: unix_daemon to instance_handler
Renames unix_daemon files to instance_handler.
Separates unix and Windows implementations of LockFileHandler
