+++ src/agent/src/windows/instance_handler_win.cpp
@@ -1,0 +1,5 @@
+#include <instance_handler.hpp>
+
+namespace unix_daemon
+{
+} // namespace unix_daemon
