+++ src/agent/tests/unix/unix_daemon_test.cpp
@@ -1,6 +1,6 @@
 #include <gtest/gtest.h>
 
-#include <unix_daemon.hpp>
+#include <instance_handler.hpp>
 
 #include <filesystem>
 #include <fstream>
